cppcheckexecutor.cpp: extracted `USE_WINDOWS_SEH` code to separate file
