+++ cli/cli.vcxproj
@@ -407,6 +407,7 @@
     <ClInclude Include="..\lib\config.h" />
     <ClInclude Include="cmdlineparser.h" />
     <ClInclude Include="cppcheckexecutor.h" />
+    <ClCompile Include="cppcheckexecutorseh.h" />
     <ClInclude Include="executor.h" />
     <ClInclude Include="filelister.h" />
     <ClInclude Include="processexecutor.h" />
@@ -420,6 +421,7 @@
   <ItemGroup>
     <ClCompile Include="cmdlineparser.cpp" />
     <ClCompile Include="cppcheckexecutor.cpp" />
+    <ClCompile Include="cppcheckexecutorseh.cpp" />
     <ClCompile Include="executor.cpp" />
     <ClCompile Include="filelister.cpp" />
     <ClCompile Include="main.cpp" />
