+++ cli/cli.vcxproj.filters
@@ -26,6 +26,9 @@
     <ClInclude Include="cppcheckexecutor.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="cppcheckexecutorseh.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="filelister.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -38,6 +41,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="cppcheckexecutor.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="cppcheckexecutorseh.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="filelister.cpp">
