+++ lib/utils.h
@@ -156,4 +156,14 @@ CPPCHECKLIB bool matchglobs(const std::vector<std::string> &patterns, const std:
 
 CPPCHECKLIB void strTolower(std::string& str);
 
+/**
+ *  Simple helper function:
+ * \return size of array
+ * */
+template<typename T, int size>
+std::size_t getArrayLength(const T (&)[size])
+{
+    return size;
+}
+
 #endif
